parse_data_new_for_seq2seq: parsing time values as float, hparams: adding new hyperparameters

+++ pp_seq2seq/data/parse_data_new_for_seq2seq.py
@@ -422,7 +422,7 @@ def main():
     data_path = dataset_path
 
     data = pd.read_csv(data_path, usecols= [0, 1], delimiter=' ', header=None, names=['ID', 'Time']) 
-    data['Time'] = data['Time'].astype(int)
+    data['Time'] = data['Time'].astype(float)
     data = data.sort_values(['Time'], ascending=True)
 
     lst = []
